requirements.txt
Django==2.1.5
psycopg2-binary==2.8.6
djangorestframework